+++ apmidproject.h
@@ -4,11 +4,11 @@
 #include <iostream>
 #include <vector>
 
-bool in_vec(const std::vector<size_t> &v, size_t a);
+bool in_vec(const std::vector<int> &v, int a);
 void hello_game();
-std::vector<size_t> own_8_puzzle_vs_random();
-std::vector<size_t> input_goal_vs_default_goal();
-void choose_algorithm(std::vector<size_t> &init, const std::vector<size_t> &goal);
+std::vector<int> own_8_puzzle_vs_random();
+std::vector<int> input_goal_vs_default_goal();
+void choose_algorithm(std::vector<int> &init, const std::vector<int> &goal);
 bool continue_game();
 
 #endif
